src/requirements.txt
django
django-sekizai
django-search-listview
